+++ engine/gears/utils.h
@@ -13,7 +13,7 @@ double random_unit();
 int32_t parse_int(const char *str, bool &ok);
 int32_t parse_int(const char *str, bool &ok, int min, int max);
 float parse_float(const char *str, bool &ok);
-float parse_float(const char *str, bool &ok, int min, int max);
+float parse_float(const char *str, bool &ok, float min, float max);
 
 void __attribute__((noinline)) breakAnchor(const char *name);
